requirements.txt
fastapi==0.95.0
numpy==1.22.4
python_Levenshtein==0.20.9
requests==2.28.2
ruamel.base==1.0.0
torch==1.12.1
tqdm==4.64.1
wandb==0.14.2